refactor : 사용자 정의 타입 제거 및 자동으로 게임 진행
